Various improvements to the horizontal bars
