+++ hit.go
@@ -9,8 +9,6 @@ import (
 	"context"
 	"fmt"
 	"net/url"
-	"sort"
-	"strconv"
 	"strings"
 	"text/tabwriter"
 	"time"
@@ -22,14 +20,6 @@ import (
 )
 
 func ptr(s string) *string { return &s }
-
-// ref_scheme column
-var (
-	RefSchemeHTTP      = ptr("h")
-	RefSchemeOther     = ptr("o")
-	RefSchemeGenerated = ptr("g")
-	RefSchemeCampaign  = ptr("c")
-)
 
 type Hit struct {
 	ID      int64  `db:"id" json:"-"`
@@ -52,143 +42,6 @@ type Hit struct {
 
 	RefURL *url.URL `db:"-" json:"-"`   // Parsed Ref
 	Random string   `db:"-" json:"rnd"` // Browser cache buster, as they don't always listen to Cache-Control
-}
-
-var groups = map[string]string{
-	// HN has <meta name="referrer" content="origin"> so we only get the domain.
-	"news.ycombinator.com":               "Hacker News",
-	"hn.algolia.com":                     "Hacker News",
-	"hckrnews.com":                       "Hacker News",
-	"hn.premii.com":                      "Hacker News",
-	"com.stefandekanski.hackernews.free": "Hacker News",
-	"io.github.hidroh.materialistic":     "Hacker News",
-	"hackerweb.app":                      "Hacker News",
-	"www.daemonology.net/hn-daily":       "Hacker News",
-	"quiethn.com":                        "Hacker News",
-	"hnews.xyz":                          "Hacker News",
-	"hackernewsmobile.com":               "Hacker News",
-	// http://www.elegantreader.com/item/17358103
-	// https://www.daemonology.net/hn-daily/2019-05.html
-
-	"mail.google.com":       "Email",
-	"com.google.android.gm": "Email",
-	"mail.yahoo.com":        "Email",
-	//  https://mailchi.mp
-
-	"org.fox.ttrss":            "RSS",
-	"www.inoreader.com":        "RSS",
-	"com.innologica.inoreader": "RSS",
-	"usepanda.com":             "RSS",
-	"feedly.com":               "RSS",
-
-	"com.google.android.googlequicksearchbox":                      "Google",
-	"com.google.android.googlequicksearchbox/https/www.google.com": "Google",
-
-	"com.andrewshu.android.reddit":       "www.reddit.com",
-	"com.laurencedawson.reddit_sync":     "www.reddit.com",
-	"com.laurencedawson.reddit_sync.dev": "www.reddit.com",
-	"com.laurencedawson.reddit_sync.pro": "www.reddit.com",
-
-	"m.facebook.com":  "www.facebook.com",
-	"l.facebook.com":  "www.facebook.com",
-	"lm.facebook.com": "www.facebook.com",
-
-	"org.telegram.messenger": "Telegram Messenger",
-
-	"com.Slack": "Slack Chat",
-}
-
-var hostAlias = map[string]string{
-	"en.m.wikipedia.org": "en.wikipedia.org",
-	"m.facebook.com":     "www.facebook.com",
-	"m.habr.com":         "habr.com",
-	"old.reddit.com":     "www.reddit.com",
-	"i.reddit.com":       "www.reddit.com",
-	"np.reddit.com":      "www.reddit.com",
-	"fr.reddit.com":      "www.reddit.com",
-}
-
-func cleanRefURL(ref string, refURL *url.URL) (string, bool) {
-	// I'm not sure where these links are generated, but there are *a lot* of
-	// them.
-	if refURL.Host == "link.oreilly.com" {
-		return "link.oreilly.com", false
-	}
-
-	// Always remove protocol.
-	refURL.Scheme = ""
-	if p := strings.Index(ref, ":"); p > -1 && p < 7 {
-		ref = ref[p+3:]
-	}
-
-	// Normalize some hosts.
-	if a, ok := hostAlias[refURL.Host]; ok {
-		refURL.Host = a
-	}
-
-	// Group based on URL.
-	if strings.HasPrefix(refURL.Host, "www.google.") {
-		// Group all "google.co.nz", "google.nl", etc. as "Google".
-		return "Google", true
-	}
-	if g, ok := groups[refURL.Host]; ok {
-		return g, true
-	}
-
-	// Useful: https://lobste.rs/s/tslw6k/why_i_m_still_using_jquery_2019
-	// Not really: https://lobste.rs/newest/page/8, https://lobste.rs/page/7
-	//             https://lobste.rs/search, https://lobste.rs/t/javascript
-	if refURL.Host == "lobste.rs" && !strings.HasPrefix(refURL.Path, "/s/") {
-		return "lobste.rs", false
-	}
-	if refURL.Host == "gambe.ro" && !strings.HasPrefix(refURL.Path, "/s/") {
-		return "lobste.rs", false
-	}
-
-	// Reddit
-	// www.reddit.com/r/programming/top
-	// www.reddit.com/r/programming/.compact
-	// www.reddit.com/r/programming.compact
-	// www.reddit.com/r/webdev/new
-	// www.reddit.com/r/vim/search
-	if refURL.Host == "www.reddit.com" {
-		switch {
-		case strings.HasSuffix(refURL.Path, "/top") || strings.HasSuffix(refURL.Path, "/new"):
-			refURL.Path = refURL.Path[:len(refURL.Path)-4]
-		case strings.HasSuffix(refURL.Path, "/search"):
-			refURL.Path = refURL.Path[:len(refURL.Path)-7]
-		case strings.HasSuffix(refURL.Path, ".compact"):
-			refURL.Path = refURL.Path[:len(refURL.Path)-8]
-		}
-	}
-
-	// Linking https://t.co/c3MITw38Yq isn't too useful as that will link back
-	// to the page, so link to the Tweet instead.
-	if refURL.Host == "t.co" {
-		return "twitter.com/search?q=https%3A%2F%2Ft.co" +
-			url.QueryEscape(refURL.Path), false
-	}
-
-	// Clean query parameters.
-	i := strings.Index(ref, "?")
-	if i == -1 {
-		// No parameters so no work.
-		return strings.TrimLeft(refURL.String(), "/"), false
-	}
-
-	q := refURL.Query()
-	refURL.RawQuery = ""
-
-	// Google analytics tracking parameters.
-	q.Del("utm_source")
-	q.Del("utm_medium")
-	q.Del("utm_campaign")
-	q.Del("utm_term")
-
-	if len(q) == 0 {
-		return refURL.String()[2:], false
-	}
-	return refURL.String()[2:], false
 }
 
 func (h *Hit) cleanPath(ctx context.Context) {
@@ -499,251 +352,4 @@ func (h *Stats) ByRef(ctx context.Context, start, end time.Time, ref string, lim
 		total += b.Count
 	}
 	return total, errors.Wrap(err, "HitStats.ByRef")
-}
-
-// ListBrowsers lists all browser statistics for the given time period.
-func (h *Stats) ListBrowsers(ctx context.Context, start, end time.Time) (int, error) {
-	err := zdb.MustGet(ctx).SelectContext(ctx, h, `/* Stats.ListBrowsers */
-		select
-			browser as name,
-			sum(count) as count,
-			sum(count_unique) as count_unique
-		from browser_stats
-		where site=$1 and day >= $2 and day <= $3
-		group by browser
-		order by count desc
-	`, MustGetSite(ctx).ID, start.Format("2006-01-02"), end.Format("2006-01-02"))
-	if err != nil {
-		return 0, errors.Wrap(err, "Stats.ListBrowsers browsers")
-	}
-
-	var total int
-	for _, b := range *h {
-		total += b.Count
-	}
-
-	return total, nil
-}
-
-// ListBrowser lists all the versions for one browser.
-func (h *Stats) ListBrowser(ctx context.Context, browser string, start, end time.Time) (int, error) {
-	err := zdb.MustGet(ctx).SelectContext(ctx, h, `
-		select
-			browser || ' ' || version as name,
-			sum(count) as count,
-			sum(count_unique) as count_unique
-		from browser_stats
-		where site=$1 and day >= $2 and day <= $3 and lower(browser)=lower($4)
-		group by browser, version
-		order by count desc
-	`, MustGetSite(ctx).ID, start.Format("2006-01-02"), end.Format("2006-01-02"), browser)
-	if err != nil {
-		return 0, errors.Wrap(err, "Stats.ListBrowser")
-	}
-
-	var total int
-	for _, b := range *h {
-		total += b.Count
-	}
-	return total, nil
-}
-
-// ListSystems lists OS statistics for the given time period.
-func (h *Stats) ListSystems(ctx context.Context, start, end time.Time) (int, error) {
-	err := zdb.MustGet(ctx).SelectContext(ctx, h, `/* Stats.ListSystem */
-		select
-			system as name,
-			sum(count) as count,
-			sum(count_unique) as count_unique
-		from system_stats
-		where site=$1 and day >= $2 and day <= $3
-		group by system
-		order by count desc
-	`, MustGetSite(ctx).ID, start.Format("2006-01-02"), end.Format("2006-01-02"))
-	if err != nil {
-		return 0, errors.Wrap(err, "Stats.OS")
-	}
-
-	var total int
-	for _, b := range *h {
-		total += b.Count
-	}
-
-	return total, nil
-}
-
-// ListSystem lists all the versions for one system.
-func (h *Stats) ListSystem(ctx context.Context, system string, start, end time.Time) (int, error) {
-	err := zdb.MustGet(ctx).SelectContext(ctx, h, `
-		select
-			system || ' ' || version as name,
-			sum(count) as count,
-			sum(count_unique) as count_unique
-		from system_stats
-		where site=$1 and day >= $2 and day <= $3 and lower(system)=lower($4)
-		group by system, version
-		order by count desc
-	`, MustGetSite(ctx).ID, start.Format("2006-01-02"), end.Format("2006-01-02"), system)
-	if err != nil {
-		return 0, errors.Wrap(err, "Stats.ListSystem")
-	}
-
-	var total int
-	for _, b := range *h {
-		total += b.Count
-	}
-	return total, nil
-}
-
-const (
-	sizePhones      = "Phones"
-	sizeLargePhones = "Large phones, small tablets"
-	sizeTablets     = "Tablets and small laptops"
-	sizeDesktop     = "Computer monitors"
-	sizeDesktopHD   = "Computer monitors larger than HD"
-	sizeUnknown     = "(unknown)"
-)
-
-// ListSizes lists all device sizes.
-func (h *Stats) ListSizes(ctx context.Context, start, end time.Time) (int, error) {
-	err := zdb.MustGet(ctx).SelectContext(ctx, h, `/* Stats.ListSizes */
-		select
-			width as name,
-			sum(count) as count,
-			sum(count_unique) as count_unique
-		from size_stats
-		where site=$1 and day >= $2 and day <= $3
-		group by width
-		order by count desc
-	`, MustGetSite(ctx).ID, start.Format("2006-01-02"), end.Format("2006-01-02"))
-	if err != nil {
-		return 0, errors.Wrap(err, "Stats.ListSize")
-	}
-
-	// Group a bit more user-friendly.
-	// TODO: ideally I'd like to make a line chart in the future, in which case
-	// this should no longer be needed.
-	ns := Stats{
-		{Name: sizePhones, Count: 0, CountUnique: 0},
-		{Name: sizeLargePhones, Count: 0, CountUnique: 0},
-		{Name: sizeTablets, Count: 0, CountUnique: 0},
-		{Name: sizeDesktop, Count: 0, CountUnique: 0},
-		{Name: sizeDesktopHD, Count: 0, CountUnique: 0},
-		{Name: sizeUnknown, Count: 0, CountUnique: 0},
-	}
-
-	hh := *h
-	var count int
-	for i := range hh {
-		count += hh[i].Count
-
-		x, _ := strconv.ParseInt(hh[i].Name, 10, 16)
-		switch {
-		case x == 0:
-			ns[5].Count += hh[i].Count
-			ns[5].CountUnique += hh[i].CountUnique
-		case x <= 384:
-			ns[0].Count += hh[i].Count
-			ns[0].CountUnique += hh[i].CountUnique
-		case x <= 1024:
-			ns[1].Count += hh[i].Count
-			ns[1].CountUnique += hh[i].CountUnique
-		case x <= 1440:
-			ns[2].Count += hh[i].Count
-			ns[2].CountUnique += hh[i].CountUnique
-		case x <= 1920:
-			ns[3].Count += hh[i].Count
-			ns[3].CountUnique += hh[i].CountUnique
-		default:
-			ns[4].Count += hh[i].Count
-			ns[4].CountUnique += hh[i].CountUnique
-		}
-	}
-	*h = ns
-
-	return count, nil
-}
-
-// ListSize lists all sizes for one grouping.
-func (h *Stats) ListSize(ctx context.Context, name string, start, end time.Time) (int, error) {
-	var where string
-	switch name {
-	case sizePhones:
-		where = "width != 0 and width <= 384"
-	case sizeLargePhones:
-		where = "width != 0 and width <= 1024 and width > 384"
-	case sizeTablets:
-		where = "width != 0 and width <= 1440 and width > 1024"
-	case sizeDesktop:
-		where = "width != 0 and width <= 1920 and width > 1440"
-	case sizeDesktopHD:
-		where = "width != 0 and width > 1920"
-	case sizeUnknown:
-		where = "width = 0"
-	default:
-		return 0, errors.Errorf("Stats.ListSizes: invalid value for name: %#v", name)
-	}
-
-	err := zdb.MustGet(ctx).SelectContext(ctx, h, fmt.Sprintf(`/* Stats.ListLocations */
-		select
-			width as name,
-			sum(count) as count,
-			sum(count_unique) as count_unique
-		from size_stats
-		where
-			site=$1 and day >= $2 and day <= $3 and
-			%s
-		group by width
-	`, where), MustGetSite(ctx).ID, start.Format("2006-01-02"), end.Format("2006-01-02"))
-	if err != nil {
-		return 0, errors.Wrap(err, "Stats.ListSize")
-	}
-
-	grouped := make(map[string]int)
-	groupedUnique := make(map[string]int)
-	hh := *h
-	for i := range hh {
-		grouped[fmt.Sprintf("↔ %spx", hh[i].Name)] += hh[i].Count
-		groupedUnique[fmt.Sprintf("↔ %spx", hh[i].Name)] += hh[i].CountUnique
-	}
-
-	ns := Stats{}
-	total := 0
-	for width, count := range grouped {
-		total += count
-		ns = append(ns, StatT{
-			Name:        width,
-			Count:       count,
-			CountUnique: groupedUnique[width],
-		})
-	}
-	sort.Slice(ns, func(i int, j int) bool { return ns[i].Count > ns[j].Count })
-	*h = ns
-
-	return total, nil
-}
-
-// ListLocations lists all location statistics for the given time period.
-func (h *Stats) ListLocations(ctx context.Context, start, end time.Time) (int, error) {
-	err := zdb.MustGet(ctx).SelectContext(ctx, h, `/* Stats.ListLocations */
-		select
-			iso_3166_1.name as name,
-			sum(count) as count,
-			sum(count_unique) as count_unique
-		from location_stats
-		join iso_3166_1 on iso_3166_1.alpha2=location
-		where site=$1 and day >= $2 and day <= $3
-		group by location, iso_3166_1.name
-		order by count desc
-	`, MustGetSite(ctx).ID, start.Format("2006-01-02"), end.Format("2006-01-02"))
-	if err != nil {
-		return 0, errors.Wrap(err, "Stats.ListLocations")
-	}
-
-	var total int
-	for _, b := range *h {
-		total += b.Count
-	}
-
-	return total, nil
 }
