+++ public/style_backend.css
@@ -307,13 +307,14 @@ select#timezone { max-width: 20rem; }
 
 table.auto { width: auto; }
 
-.browser-charts          { display: flex; flex-wrap: wrap; justify-content: space-between; }
-.browser-charts > div    { width: 49%; }
-.browser-charts h2 small { float: right; font-variant-ligatures: none; font-feature-settings: 'liga' off, 'dlig' off; }
+.hcharts            { display: flex; flex-wrap: wrap; justify-content: space-between; }
+.hcharts > div      { width: 49%; }
+.hcharts h2 small   { float: right; font-variant-ligatures: none; font-feature-settings: 'liga' off, 'dlig' off; }
+.hcharts .load-more { margin-top: .5em; }
 
 @media (max-width: 45rem) {
-	.browser-charts       { display: block; }
-	.browser-charts > div { width: auto; }
+	.hcharts       { display: block; }
+	.hcharts > div { width: auto; }
 }
 
 .chart-hbar a, .chart-hbar > p {
@@ -344,10 +345,11 @@ table.auto { width: auto; }
 .chart-hbar.hbar-detail > *            { cursor: default; }
 .chart-hbar.hbar-detail > *:hover span { background-color: #9a15a4;  }
 .chart-hbar.hbar-detail > *:focus      { outline: none; }
-.chart-hbar > *[title^="(other): "],       .chart-hbar > *[title^="(unknown): "]       { cursor: default; font-style: italic; }
-.chart-hbar > *[title^="(other): "]:hover, .chart-hbar > *[title^="(unknown): "]:hover { color: #252525; }
-.chart-hbar > *[title^="(other): "]:hover span, .chart-hbar > *[title^="(unknown): "]:hover span { background-color: #9a15a4; }
-.chart-hbar > *[title^="(other): "]:focus, .chart-hbar > *[title^="(unknown): "]:focus { outline: none; }
+
+.chart-hbar .other            { cursor: default; font-style: italic; }
+.chart-hbar .other:hover      { color: #252525; }
+.chart-hbar .other:hover span { background-color: #9a15a4; }
+.chart-hbar .other:focus      { outline: none; }
 
 .hchart-wrap { position: relative; }
 .hbar-open   { opacity: .25; background-color: #ddd; }
